+++ daemon/src/config_old.rs
@@ -25,6 +25,7 @@ impl ConfigV212 {
             gfx_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
             active_profile: self.active_profile,
+            gfx_vfio_enable: false,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.power_profile,
             bat_charge_limit: self.bat_charge_limit,
@@ -53,6 +54,7 @@ impl ConfigV222 {
         Config {
             gfx_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
+            gfx_vfio_enable: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.power_profile,
@@ -84,6 +86,7 @@ impl ConfigV301 {
         Config {
             gfx_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
+            gfx_vfio_enable: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.curr_fan_mode,
@@ -115,6 +118,34 @@ impl ConfigV317 {
         Config {
             gfx_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
+            gfx_vfio_enable: false,
+            active_profile: self.active_profile,
+            toggle_profiles: self.toggle_profiles,
+            curr_fan_mode: self.curr_fan_mode,
+            bat_charge_limit: self.bat_charge_limit,
+            power_profiles: self.power_profiles,
+        }
+    }
+}
+
+#[derive(Deserialize, Serialize)]
+pub struct ConfigV324 {
+    pub gfx_mode: GfxVendors,
+    pub gfx_managed: bool,
+    pub active_profile: String,
+    pub toggle_profiles: Vec<String>,
+    #[serde(skip)]
+    pub curr_fan_mode: u8,
+    pub bat_charge_limit: u8,
+    pub power_profiles: BTreeMap<String, Profile>,
+}
+
+impl ConfigV324 {
+    pub(crate) fn into_current(self) -> Config {
+        Config {
+            gfx_mode: GfxVendors::Hybrid,
+            gfx_managed: self.gfx_managed,
+            gfx_vfio_enable: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.curr_fan_mode,